rearrange functions in proper files
