+++ minihell/parsing/free.c
@@ -1,35 +1,35 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   env.c                                              :+:      :+:    :+:   */
+/*   free.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: abenajib <abenajib@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/05 13:21:14 by abenajib          #+#    #+#             */
-/*   Updated: 2025/04/08 11:06:55 by abenajib         ###   ########.fr       */
+/*   Created: 2025/04/08 11:00:00 by abenajib          #+#    #+#             */
+/*   Updated: 2025/04/08 11:01:03 by abenajib         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-t_list	*ft_envinit(char **env)
+void	ft_free_tokenlist(t_token *token_list)
 {
-	t_list	*minienv;
-	char	*key;
-	char	*value;
-	int		i;
+	t_token	*tmp;
 
-	minienv = NULL;
-	i = 0;
-	while (env[i])
+	while (token_list)
 	{
-		key = ft_substr(env[i], 0, ft_strchr(env[i], '=') - env[i]);
-		value = ft_substr(env[i], ft_strchr(env[i], '=') - env[i] + 1,
-				ft_strlen(env[i]) - (ft_strchr(env[i], '=') - env[i]));
-		if (!key || !value)
-			return (NULL);
-		ft_lstadd_back(&minienv, ft_lstnew(key, value));
-		i++;
+		tmp = token_list;
+		token_list = token_list->next;
+		free(tmp->value);
+		free(tmp);
 	}
-	return (minienv);
+}
+
+void	free_token(t_token *token)
+{
+	if (token)
+	{
+		free(token->value);
+		free(token);
+	}
 }
