+++ minihell/parsing/tokens.c
@@ -1,0 +1,63 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   tokens.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: abenajib <abenajib@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/08 11:02:14 by abenajib          #+#    #+#             */
+/*   Updated: 2025/04/08 11:04:09 by abenajib         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../minishell.h"
+
+void	ft_tokadd_back(t_token **token_list, t_token *token)
+{
+	t_token	*tmp;
+
+	if (!token_list || !token)
+		return ;
+	if (*token_list == NULL)
+	{
+		*token_list = token;
+		return ;
+	}
+	tmp = *token_list;
+	while (tmp->next)
+		tmp = tmp->next;
+	tmp->next = token;
+	token->next = NULL;
+}
+
+t_token	*ft_newtok(t_token *token)
+{
+	t_token	*new_token;
+
+	new_token = malloc(sizeof(t_token));
+	if (!new_token)
+		return (NULL);
+	new_token->type = token->type;
+	new_token->value = ft_strdup(token->value);
+	new_token->quote_type = token->quote_type;
+	new_token->next = NULL;
+	return (new_token);
+}
+
+t_token	*get_next_token(t_lexer *lexer)
+{
+	char	current_char;
+
+	while (lexer->pos < lexer->len
+		&& ft_isspace(lexer->input[lexer->pos]))
+		lexer->pos++;
+	if (lexer->pos >= lexer->len)
+		return (NULL);
+	current_char = lexer->input[lexer->pos];
+	if (current_char == '\'' || current_char == '"')
+		return (ft_handle_quotes(lexer, current_char));
+	else if (ft_isspecial(current_char))
+		return (ft_handle_operator(lexer));
+	else
+		return (ft_handle_word(lexer));
+}
