+++ minihell/parsing/utils.c
@@ -6,7 +6,7 @@
 /*   By: abenajib <abenajib@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/07 18:10:47 by abenajib          #+#    #+#             */
-/*   Updated: 2025/04/07 20:11:38 by abenajib         ###   ########.fr       */
+/*   Updated: 2025/04/08 11:00:54 by abenajib         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,13 +30,4 @@ bool	ft_is_duplicated(t_lexer *lexer, char op[3])
 			&& lexer->input[lexer->pos + 1] == '>')
 		|| (op[0] == '<' && lexer->pos + 1 < lexer->len
 			&& lexer->input[lexer->pos + 1] == '<'));
-}
-
-void	free_token(t_token *token)
-{
-	if (token)
-	{
-		free(token->value);
-		free(token);
-	}
 }
